+++ frontend/src/components/SignUpLink/SignUpLink.css
@@ -1,2 +1,5 @@
-
-
+.sign-up-link p.register {
+    margin-left: 5px;
+    color: blue;
+    font-style: italic;
+}
