+++ frontend/src/views/CoinView/CoinView.js
@@ -2,50 +2,36 @@ import React,{useState, useContext, useRef} from "react";
 import './CoinView.css'
 import CoinHistoryGraph from '../../components/CoinHistoryGraph/CoinHistoryGraph'
 import { CoinInfoBox } from '../../components/CoinInfoBox/CoinInfoBox'
-import Forecast from '../../components/Forecast/Forecast'
+
 import SaveLastViewed from '../../components/SaveLastViewed/SaveLastViewed'
 import CryptoShuttleService from "../../utils/api/services/CryptoShuttleService";
 import { UserContext } from '../../shared/global/provider/UserProvider'
 import { SignUpLink } from "../../components/SignUpLink/SignUpLink";
-
-
+import { LoginButton } from "../../components/LoginButton/LoginButton";
+import { RunForecastButton } from "../../components/Forecast/RunForecastButton";
 
 
 export const CoinView = () => {
-    let btnRef = useRef();
-    const [showResults, setShowResults] = React.useState(false)
+    const signInRef = useRef()
     const [authenticatedUser, setAuthenticatedUser] = useContext(UserContext)
-    let  [user, setUser] = useState()
-    const onClick = () => {setShowResults(true)
-        
-        if(btnRef.current){
-            btnRef.current.setAttribute("hidden", "hidden");
-        }
-        userAuthBackend()
-        
-    }
-        
-            
-    const userAuthBackend = async () => {
-        
-        setAuthenticatedUser(localStorage.getItem('token'))
-        console.log(authenticatedUser)
 
-        if(authenticatedUser == null){
-            console.log('login to use')
-        }else if(authenticatedUser != null) {
-                console.log(authenticatedUser)
-            try{
-                const getUserWithToken = await CryptoShuttleService.getLoggedInUser(localStorage.getItem('token'))
-                setUser(getUserWithToken.data.logged_in_as)
-                
-            }
-            catch(e){
-                console.log(e.message);
-            }
-        }
+    
+    
+    const displayIfNotAuth = () => {
         
+        
+        if(authenticatedUser != null && signInRef.current) {
+            signInRef.current.setAttribute("hidden", "hidden")
+        }
     }
+    
+    
+    
+    
+    
+    
+       
+    
     
       
     return (
@@ -55,10 +41,16 @@ export const CoinView = () => {
                 <CoinHistoryGraph />
                 <SaveLastViewed />
                 
-                <input ref={btnRef} type="submit" value="Run Forecast" onClick={onClick} />
-                { showResults && user && <Forecast /> }
-                { !user && showResults && <SignUpLink/> }
+                <RunForecastButton/>
+
+                <div ref={signInRef} className="sign-in">
+                    {authenticatedUser == null && <LoginButton/>}
+                    {authenticatedUser == null && <SignUpLink/>}
+                    {displayIfNotAuth()}
+                </div>
                 
+               
+               
                                
             </div>
         </div>
